+++ pypeit/tests/test_match.py
@@ -7,6 +7,7 @@ import pytest
 
 import numpy as np
 
+from pypeit.calibframe import CalibFrame
 from pypeit.core import framematch
 from pypeit.tests.tstutils import dummy_fitstbl
 from pypeit.pypmsgs import PypeItError
@@ -62,17 +63,19 @@ def test_calibration_groups(fitstbl):
 
 
 def test_instr_setup(fitstbl):
-    """ Test instrument setup naming convention
-    Tickles most of the arsetup methods
     """
-    par = fitstbl.spectrograph.default_pypeit_par()
+    Test instrument setup naming convention
+    """
+    detname = fitstbl.spectrograph.get_det_name(1)
+    calib_key = CalibFrame.construct_calib_key(fitstbl['setup'][0], fitstbl['calib'][0], detname)
 
-    # Check the master key
-    assert fitstbl.master_key(0) == 'A_1_DET01'
+    # Check the calibration  key
+    assert calib_key == 'A_0_DET01'
     # Invalid detector
     with pytest.raises(PypeItError):
         # Shane kast blue doesn't have a second detector
-        fitstbl.master_key(0, det=2)
+        fitstbl.spectrograph.get_det_name(2)
+
 
 
 # TODO: Need a test that adds a calibration group and checks the result
@@ -100,5 +103,20 @@ def test_exptime():
                           np.array([True, True, False, False]))
     assert np.array_equal(framematch.check_frame_exptime(exptime, [10,20]),
                           np.array([False, False, False, False]))
+
+
+def test_valid_frametype():
+
+    assert framematch.valid_frametype('arc'), 'arc should be a valid frametype'
+    assert framematch.valid_frametype('arc', quiet=True), 'arc should be a valid frametype'
+    assert not framematch.valid_frametype('junk', quiet=True, raise_error=False), \
+            'junk should not be a valid frametype'
+    assert not framematch.valid_frametype('junk', quiet=False, raise_error=False), \
+            'junk should not be a valid frametype'
+    # These should raise errors
+    with pytest.raises(PypeItError):
+        framematch.valid_frametype('junk', quiet=False, raise_error=True)
+    with pytest.raises(PypeItError):
+        framematch.valid_frametype('junk', quiet=True, raise_error=True)